+++ src/app/constant/apiConstants.ts
@@ -1,0 +1,6 @@
+export const TOUR_API_KEY = process.env.NEXT_PUBLIC_TOUR_API_KEY;
+
+export const API_URLS = {
+  TOUR_LIST: 'https://apis.data.go.kr/B551011/KorService1/areaBasedList1',
+  LOCATION_BASED: 'https://apis.data.go.kr/B551011/KorService1/locationBasedList1',
+};
